feat: extract ExecutionService from AutoModeService
Move executeFeature and all helper methods into dedicated ExecutionService
class in apps/server/src/services/auto-mode/execution-service.ts.

Extracted methods: runAgent, buildTaskPrompt, executePipelineSteps,
buildPipelineStepPrompt, getPlanningPromptPrefix, getHeapUsagePercent,
recordLearningsFromFeature (~2,852 lines removed from auto-mode-service.ts).

AutoModeService now delegates:
- executeFeature → executionService.executeFeature()
- resumeFromPipelineStep → executionService.executePipelineSteps()
- followUpFeature → executionService.runAgent()

Execution types extracted to execution-types.ts: IAutoModeCallbacks,
RunningFeature, PendingApproval, ParsedTask, PlanSpec, ExecuteFeatureOptions.
